+++ FilterTheSpire2/Code/Ancients/Filtering/Tezcatara.cs
@@ -10,7 +10,7 @@ public class Tezcatara : AbstractAncient
         Id = "TEZCATARA";
     }
 
-    public override bool CheckOptions(uint seed, Enum? relicOption)
+    public override bool CheckOptions(ulong seed, Enum? relicOption)
     {
         if (relicOption is not TezcataraOptions relic)
         {
