+++ FilterTheSpire2/Code/Filters/PotionFilters/PotionSource.cs
@@ -1,3 +1,4 @@
+using FilterTheSpire2.Code.Helpers;
 using MegaCrit.Sts2.Core.Random;
 
 namespace FilterTheSpire2.Code.Filters.PotionFilters;
@@ -6,9 +7,9 @@ public abstract class PotionSource(int priorConsumptionSteps)
 {
     public abstract int Count { get; }
 
-    protected abstract Rng GetBaseRng(uint seed);
+    protected abstract Rng GetBaseRng(ulong seed);
 
-    public Rng GetRng(uint seed)
+    public Rng GetRng(ulong seed)
     {
         var rng = GetBaseRng(seed);
         if (priorConsumptionSteps > 0)
