+++ FilterTheSpire2/Code/Filters/RelicOutcomeFilters/KaleidoscopeFilter.cs
@@ -39,8 +39,8 @@ public class KaleidoscopeFilter : BaseCardRewardFilter
     }
 
     protected override bool IsCharacterRequired => true;
-    protected override Rng GetRewardRng(uint seed) => RngHelper.GetPlayerRngType(seed, PlayerRngType.Rewards);
-    protected override Rng? GetCardPoolRng(uint seed) => RngHelper.GetRunRngType(seed, RunRngType.Niche);
+    protected override Rng GetRewardRng(ulong seed) => RngHelper.GetPlayerRngType(seed, PlayerRngType.Rewards);
+    protected override Rng? GetCardPoolRng(ulong seed) => RngHelper.GetRunRngType(seed, RunRngType.Niche);
     protected override List<List<CardDefinition>> GetRewardPools(Rng rng)
     {
         var charList = _charListFiltered.ToList();
